fixed issue in SLL
